+++ objetos_celestes.hpp
@@ -2,23 +2,24 @@
 #define OBJETOS_CELESTES_HPP_INCLUDED
 //||\n\033[0m?
 #include <string>
+#include <cmath>
 
-float calculo_distancia_pela_esfera_de_Hill(float Raio_de_Hill, float excentricidade, float massa_menor, float Massa_maior);
+double calculo_distancia_pela_esfera_de_Hill(double Raio_de_Hill, double excentricidade, double massa_menor, double Massa_maior);
 
 struct Orbita{
-    float apoastro = 0.0;
-    float periastro = 0.0;
-    float excentricidade = 0.0f;
+    double apoastro = 0.0;
+    double periastro = 0.0;
+    double excentricidade = 0.0f;
 
     Orbita();
-    Orbita(const float &apo_ou_peri, const float &excentri, const bool calcular_periastro);
+    Orbita(const double &apo_ou_peri, const double &excentri, const bool calcular_periastro);
 };
 
-Orbita calcula_orbita(float apoastro_ou_periastro, float excentricidade, bool calcular_periastro);
+Orbita calcula_orbita(double apoastro_ou_periastro, double excentricidade, bool calcular_periastro);
 
 class Estrela{
 public:
-    Estrela(const std::string &Nome,const std::string &tip,const float &Massa,const float &raio,const int &Temp);
+    Estrela(const std::string &Nome,const std::string &tip,const double &Massa,const double &raio,const int &Temp);
     
     ~Estrela();
 
@@ -28,21 +29,21 @@ public:
     std::string get_nome()const;
     std::string get_tipo()const;
     Orbita get_orbita()const;
-    float get_massa()const;
-    float get_raio()const;
+    double get_massa()const;
+    double get_raio()const;
     int get_temperatura()const;
     
      
 private:
     
-    void define_massa(float m);
-    void define_raio(float r);
+    void define_massa(double m);
+    void define_raio(double r);
     
     std::string m_nome = "";
     std::string m_tipo = "";
     Orbita orbt;
-    float m_raio_m = 0.0;
-    float m_massa_Kg = 0.0;
+    double m_raio_m = 0.0;
+    double m_massa_Kg = 0.0;
     int m_temperatura = 0;
 };
 
@@ -50,24 +51,24 @@ private:
 // em planejamento....
 class Planeta{
 public:
-    Planeta(float m, float r, char c_t, bool roc);
+    Planeta(double m, double r, char c_t, bool roc);
     ~Planeta(){};
     
     std::string nome;
 private:
-    float m_massa_Kg = 0.0;
-    float m_raio_m = 0.0;
+    double m_massa_Kg = 0.0;
+    double m_raio_m = 0.0;
     char m_classe_tamanho = ' ';
     bool m_rochoso = false;
     
-    float m_albedo = 0.0f;
-    float m_atmosfera_densidade = 0.0f;
+    double m_albedo = 0.0f;
+    double m_atmosfera_densidade = 0.0f;
     
 };
 
-Orbita cria_orbitas_aleatorias(std::mt19937 *motor, float menor_valor, float Maior_valor, bool calcula_periastro){
+Orbita cria_orbitas_aleatorias(std::mt19937 *motor, double menor_valor, double Maior_valor, bool calcula_periastro){
     std::uniform_int_distribution<int> excent(10, 60);
-    std::uniform_real_distribution<float> periastro_ou_apoastro(menor_valor, Maior_valor);
+    std::uniform_real_distribution<double> periastro_ou_apoastro(menor_valor, Maior_valor);
     Orbita orbit;
     
     if(calcula_periastro){
@@ -85,10 +86,10 @@ Orbita cria_orbitas_aleatorias(std::mt19937 *motor, float menor_valor, float Mai
     return orbit;
 }
     
-//float Calculo_Massivo_da_esfera_de_Hill(float distancia, float excentricidade, float massa_menor, float Massa_maior);
-//float Calculo_da_esfera_de_Hill(float distancia, float excentricidade, float massa_menor, float Massa_maior);
+//double Calculo_Massivo_da_esfera_de_Hill(double distancia, double excentricidade, double massa_menor, double Massa_maior);
+//double Calculo_da_esfera_de_Hill(double distancia, double excentricidade, double massa_menor, double Massa_maior);
 
-//float Cauculo_forca_gravitacional(float distancia, float , float massa2);
+//double Cauculo_forca_gravitacional(double distancia, double , double massa2);
 */
 
 #endif
